guardar booking API 80%

+++ database/migrations/2018_07_30_000208_create_additionals_table.php
@@ -16,8 +16,8 @@ class CreateAdditionalsTable extends Migration
         Schema::create('additionals', function (Blueprint $table) {
             $table->increments('id');
             $table->string('code', 10)->nullable()->default(null);
-            $table->double('price', 8, 2)->nullable()->default(null);
-            $table->float('iva', 5)->nullable()->default(null);
+            $table->double('price', 12, 2)->nullable()->default(null);
+            $table->double('iva', 12, 2)->nullable()->default(null);
 
             $table->unsignedInteger('additional_category_id')->default(1);
             $table->foreign('additional_category_id')->references('id')->on('additional_categories');
